working on gems

+++ lib/vendorize.rb
@@ -1,8 +1,5 @@
 #
-# Copies all required files that can be found to ENV['_Vendor_'] || './_vendor_'
-# TODO: absolute or relative paths
-# TODO: copy folders? Optionally?
-# TODO: skip options
+# Copies all required/loaded files that are needed to ENV['_Vendor_'] || './_vendor_'
 #
 class Vendorize < File
   DEFAULT_FOLDER = './_vendor_'
@@ -23,6 +20,7 @@ class Vendorize < File
   end
 
   def self.vendorize(wanted_file)
+    #puts "??? #{wanted_file}"
     vendorize('ubygems') if wanted_file =~ /^rubygems$/i
     $LOAD_PATH.each {|location|
       f = join(location, wanted_file)
@@ -33,15 +31,18 @@ class Vendorize < File
             puts "'#{wanted_file}' loaded from cache at #{file}"
           else
             dest = ensure_can_create_file("#{root_folder}/#{wanted_file}#{ext}")
-#cmd = "copy /y " + "\"#{file}\" \"#{dest}\" 2>&1".gsub(/\//,'\\')
-#`#{cmd}`
-            copyfile(file, dest)
-            puts "'#{wanted_file}' cached to #{dest} (from '#{file})"
+            unless exist?(dest)
+              copyfile(file, dest)
+              puts "'#{wanted_file}' cached to #{dest} (from '#{file})"
+            #else
+            #  puts "'#{wanted_file}' cached already to #{dest} (from '#{file})"
+            end
           end
           return
         end
       }
     }
+    #puts "no #{wanted_file}"
     nil
   end
 
@@ -60,29 +61,34 @@ class Vendorize < File
 end
 
 unless ENV['NO_VENDORIZE']  # Cannot do overrides within the tests...
+  
+  module Kernel
+    alias old_require require
+    def require(path)
+#puts "?? #{path}"
+      ret = old_require(path)
+      Vendorize.vendorize(path)
+      return ret
+      rescue LoadError => load_error
+        puts "require of #{path} failed (not necessarily a problem...)"
+#puts $LOAD_PATH
+        raise load_error
+    end
+    private :old_require
+    private :require
 
-  alias old_require require
-  def require(path)
-    Vendorize.vendorize(path)
-    old_require(path)
-    rescue LoadError => load_error
-      puts "failed"
-      raise load_error
+    alias old_load load
+    def load(filename, wrap = false)
+      puts("======================= load > #{filename}")
+      old_load(filename, wrap)
+      Vendorize.vendorize(path)
+      rescue LoadError => load_error
+        puts "load of #{path} failed (not necessarily a problem...)"
+        raise load_error
+    end
+    private :old_load
+    private :load
   end
-  private :old_require
-  private :require
-
-  alias old_load load
-  def load(filename, wrap = false)
-    puts("======================= load > #{filename}")
-    old_load(filename, wrap)
-    rescue LoadError => load_error
-      puts "failed"
-     raise load_error
-  end
-  private :old_load
-  private :load
-  $LOAD_PATH.unshift Vendorize.root_folder
 
   def vendor_only?
     false
